requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
langchain>=0.0.350
langchain-community>=0.0.10

# Fixed PyTorch and related packages
torch
torchvision==0.16.0
torchaudio==2.1.0
transformers==4.35.0
sentence_transformers

chromadb>=0.4.15
PyPDF2>=3.0.1
pymupdf>=1.23.0
cohere>=4.37
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.1.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
requests>=2.31.0
pydantic>=2.5.0
python-multipart>=0.0.6

# OCR dependencies
pytesseract>=0.3.10
opencv-python>=4.8.0
Pillow>=10.0.0

# Enhanced Bengali Processing
unicodedata2>=14.0.0
regex>=2023.10.3
